Kabir

+++ app/Http/Controllers/Auth/RoleController.php
@@ -120,17 +120,18 @@ class RoleController extends Controller
      * @param  \App\Models\Auth\Role  $role
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Role $role, $id = null)
+    public function update(Request $request, Role $role)
     {
         //
-        dd('hello!');
-        if (Gate::allows('role.edit')) {
-            $role::find($id)->update(['unique_key' => mt_rand(1000000000, 9999999999), 'name' => $request->name, 'slug' => str_replace(' ', '_', $request->name), 'status' => 1]);
-            RolePermission::where('role_id', $id)->delete();
+        if (Gate::allows('role.edit')) { 
+            $role->update(['unique_key' => mt_rand(1000000000, 9999999999), 'name' => $request->name, 'slug' => str_replace(' ', '_', $request->name), 'status' => 1]);
+           
+             RolePermission::where('role_id', $role->id)->delete();
+             
             foreach ($request->permissions as $permission) {
-                RolePermission::create(['unique_key' => mt_rand(1000000000, 9999999999), 'role_id' => $id, 'permission_id' => $permission]);
+                RolePermission::create(['unique_key' => mt_rand(1000000000, 9999999999), 'role_id' => $role->id, 'permission_id' => $permission]);
             }
-            return redirect(route('role.index'))->with('success', 'Role update Successfully!!!');
+            return redirect(route('admin.role.index'))->with('success', 'Role update Successfully!!!');
         } else {
             abort(403);
             return redirect('login');
